+++ util/filetypedetection.cpp
@@ -21,6 +21,7 @@
 #include "filetypedetection.h"
 
 #include <QImageReader>
+#include <QMediaFormat>
 #include <QMovie>
 
 FileTypeDetection::FileTypeDetection()
@@ -61,6 +62,57 @@ bool FileTypeDetection::isSupportedMovieFormat(const QMimeType &mimeType) const
     return false;
 }
 
+bool FileTypeDetection::isSupportedVideoFormat(const QMimeType &mimeType) const
+{
+    if (!mimeType.name().startsWith("video/"))
+    {
+        return false;
+    }
+
+    const QMap<QString, QMediaFormat::FileFormat> formatMapper {
+        { "video/x-ms-wmv", QMediaFormat::FileFormat::WMV },
+        { "video/x-msvideo", QMediaFormat::FileFormat::AVI },
+        { "video/x-matroska", QMediaFormat::FileFormat::Matroska },
+        { "video/mp4", QMediaFormat::FileFormat::MPEG4 },
+        { "video/ogg", QMediaFormat::FileFormat::Ogg },
+        { "video/quicktime", QMediaFormat::FileFormat::QuickTime },
+        { "video/webm", QMediaFormat::FileFormat::WebM }
+    };
+
+    if (!formatMapper.contains(mimeType.name()))
+    {
+        return false;
+    }
+    const QMediaFormat::FileFormat fileFormat = formatMapper.contains(mimeType.name())
+                                                    ? formatMapper[mimeType.name()]
+                                                    : QMediaFormat::FileFormat::UnspecifiedFormat;
+
+    QMediaFormat format;
+    const QList<QMediaFormat::FileFormat> supportedFileFormats = format.supportedFileFormats(QMediaFormat::ConversionMode::Decode);
+    if (supportedFileFormats.contains(fileFormat))
+    {
+        return true;
+    }
+
+    const QMap<QString, QMediaFormat::VideoCodec> codecMapper {
+        { "video/AV1", QMediaFormat::VideoCodec::AV1 },
+        { "video/H264", QMediaFormat::VideoCodec::H264 },
+        { "video/H265", QMediaFormat::VideoCodec::H265 },
+        { "video/mp4", QMediaFormat::VideoCodec::MPEG4 },
+        { "video/ogg", QMediaFormat::VideoCodec::Theora },
+        { "video/VP8", QMediaFormat::VideoCodec::VP8 },
+        { "video/VP9", QMediaFormat::VideoCodec::VP9 },
+        { "video/x-ms-wmv", QMediaFormat::VideoCodec::WMV }
+    };
+
+    const QMediaFormat::VideoCodec codec = codecMapper.contains(mimeType.name())
+                                               ? codecMapper[mimeType.name()]
+                                               : QMediaFormat::VideoCodec::Unspecified;
+
+    const QList<QMediaFormat::VideoCodec> supportedVideoCodecs = format.supportedVideoCodecs(QMediaFormat::ConversionMode::Decode);
+    return supportedVideoCodecs.contains(codec);
+}
+
 bool FileTypeDetection::isMovieFormat(const QMimeType &mimeType) const
 {
     return mimeType.name().startsWith("video/");
@@ -69,4 +121,9 @@ bool FileTypeDetection::isMovieFormat(const QMimeType &mimeType) const
 bool FileTypeDetection::isAudioFormat(const QMimeType &mimeType) const
 {
     return mimeType.name().startsWith("audio/");
+}
+
+bool FileTypeDetection::isVideoFormat(const QMimeType &mimeType) const
+{
+    return mimeType.name().startsWith("video/");
 }